+++ app/src/main/java/com/example/sujinv2px/evolvemvvm/ui/main/PostState.kt
@@ -1,0 +1,6 @@
+package com.example.sujinv2px.evolvemvvm.ui.main
+
+object PostState {
+
+    const val FETCH_REMOTE_COMPLETE = "fetch_remote_complete"
+}
