+++ app/src/main/java/com/example/sujinv2px/evolvemvvm/utils/StateManagementFactory.kt
@@ -1,0 +1,6 @@
+package com.example.sujinv2px.evolvemvvm.utils
+
+interface StateManagementFactory {
+
+    fun processResponse(response: Response?)
+}
